Extracted ListOrDict.
